requirements.txt
google-generativeai
PyPDF2
python-dotenv
tenacity